qwen/backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via typer
anyio==4.12.1
    # via httpx
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via pitstoptimingparsing (pyproject.toml)
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via typer
colorama==0.4.6
    # via
    #   click
    #   pytest
coverage==7.13.4
    # via pytest-cov
django==5.2.12
    # via
    #   pitstoptimingparsing (pyproject.toml)
    #   django-cors-headers
    #   django-filter
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.9.0
    # via pitstoptimingparsing (pyproject.toml)
django-filter==25.2
    # via pitstoptimingparsing (pyproject.toml)
django-rest-framework==0.1.0
    # via pitstoptimingparsing (pyproject.toml)
django-stubs==5.2.9
    # via djangorestframework-stubs
django-stubs-ext==5.2.9
    # via django-stubs
djangorestframework==3.16.1
    # via
    #   django-rest-framework
    #   drf-spectacular
djangorestframework-stubs==3.16.8
    # via pitstoptimingparsing (pyproject.toml)
dotenv==0.9.9
    # via pitstoptimingparsing (pyproject.toml)
drf-spectacular==0.29.0
    # via pitstoptimingparsing (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via pitstoptimingparsing (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.3.0
    # via pytest
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
lxml==6.0.2
    # via pitstoptimingparsing (pyproject.toml)
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==26.0
    # via pytest
pendulum==3.2.0
    # via pitstoptimingparsing (pyproject.toml)
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg==3.3.3
    # via pitstoptimingparsing (pyproject.toml)
psycopg2-binary==2.9.11
    # via psycopg
    # via pitstoptimingparsing (pyproject.toml)
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==9.0.2
    # via
    #   pitstoptimingparsing (pyproject.toml)
    #   pytest-cov
    #   pytest-django
pytest-cov==7.0.0
    # via pitstoptimingparsing (pyproject.toml)
pytest-django==4.12.0
    # via pitstoptimingparsing (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pendulum
python-decouple==3.8
    # via pitstoptimingparsing (pyproject.toml)
python-dotenv==1.2.1
    # via dotenv
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via pitstoptimingparsing (pyproject.toml)
rich==14.3.3
    # via
    #   pitstoptimingparsing (pyproject.toml)
    #   typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruff==0.15.2
    # via pitstoptimingparsing (pyproject.toml)
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
sqlparse==0.5.5
    # via django
typer==0.24.0
    # via pitstoptimingparsing (pyproject.toml)
types-pyyaml==6.0.12.20250915
    # via
    #   django-stubs
    #   djangorestframework-stubs
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
tzdata==2025.3
    # via
    #   django
    #   pendulum
    #   psycopg
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.6.3
    # via requests
